# ai-service/tests/api/fixtures/openapi-c40-baseline.json
{
  "components": {
    "schemas": {
...
      "DebugInfo": {
        "description": "Optional retrieval diagnostics; minimally filled by the stubs.",
        "properties": {
          "lexical_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lexical Score"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "rerank_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rerank Score"
          },
          "vector_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vector Score"
          }
        },
        "title": "DebugInfo",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetrievalFilters": {
        "description": "Catalog-side filters only. Price and stock filtering stays on the .NET side.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "exclude_product_ids": {
            "items": {
              "type": "string"
            },
            "title": "Exclude Product Ids",
            "type": "array"
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Family Id"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "title": "Materials",
            "type": "array"
          }
        },
        "title": "RetrievalFilters",
        "type": "object"
      },
      "RetrievalMode": {
        "enum": [
          "hybrid",
          "vector",
          "lexical"
        ],
        "title": "RetrievalMode",
        "type": "string"
      },
      "RetrievalRequest": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/RetrievalFilters"
          },
          "mode": {
            "$ref": "#/components/schemas/RetrievalMode",
            "default": "hybrid"
          },
          "pos_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Accepted for client compatibility and ignored; scope comes from the token",
            "title": "Pos Id"
          },
          "query": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "top_k": {
            "default": 10,
            "description": "Page size wanted after hydration",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "query"
        ],
        "title": "RetrievalRequest",
        "type": "object"
      },
      "RetrievalResponse": {
        "properties": {
          "candidates_returned": {
            "description": "Candidates the retriever produced",
            "minimum": 0.0,
            "title": "Candidates Returned",
            "type": "integer"
          },
          "effective_pos_id": {
            "description": "Scope applied by the handler; always the token claim, never the body value",
            "title": "Effective Pos Id",
            "type": "string"
          },
          "low_confidence": {
            "default": false,
            "title": "Low Confidence",
            "type": "boolean"
          },
          "projection_age_seconds": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "description": "Seconds since the point-of-sale availability projection was last synchronised. Null when the prefilter did not run. Taken from the drain checkpoint and never from the projection rows: the feed is incremental, so a row's timestamp records when that assignment last changed rather than when the projection was last read, and would report months on a projection synchronised seconds ago. Above the configured ceiling the point-of-sale scope is not applied for that request and this field is how the caller can tell.",
            "title": "Projection Age Seconds"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/RetrievalResult"
            },
            "title": "Results",
            "type": "array"
          },
          "trace_id": {
            "description": "Correlation id, from the token claim when present",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "effective_pos_id",
          "results",
          "candidates_returned"
        ],
        "title": "RetrievalResponse",
        "type": "object"
      },
      "RetrievalResult": {
        "properties": {
          "debug": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DebugInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the family is unknown",
            "title": "Family Id"
          },
          "match_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Match Reasons",
            "type": "array"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "title": "Materials",
            "type": "array"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "variant_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the variant is unknown",
            "title": "Variant Label"
          }
        },
        "required": [
          "product_id",
          "sku",
          "score"
        ],
        "title": "RetrievalResult",
        "type": "object"
      },
      "SimilaritySignals": {
        "description": "Why two products are considered interchangeable. No price signal by design.",
        "properties": {
          "family_match": {
            "default": false,
            "title": "Family Match",
            "type": "boolean"
          },
          "material_overlap": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Material Overlap",
            "type": "number"
          },
          "style_similarity": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Style Similarity",
            "type": "number"
          },
          "visual_similarity": {
            "anyOf": [
              {
                "maximum": 1.0,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Visual Similarity"
          }
        },
        "required": [
          "material_overlap",
          "style_similarity"
        ],
        "title": "SimilaritySignals",
        "type": "object"
      },
      "SubstituteResult": {
        "properties": {
          "debug": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DebugInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "family_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the family is unknown",
            "title": "Family Id"
          },
          "match_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Match Reasons",
            "type": "array"
          },
          "materials": {
            "items": {
              "type": "string"
            },
            "title": "Materials",
            "type": "array"
          },
          "product_id": {
            "title": "Product Id",
            "type": "string"
          },
          "score": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score",
            "type": "number"
          },
          "similarity_signals": {
            "$ref": "#/components/schemas/SimilaritySignals"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "variant_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null when the variant is unknown",
            "title": "Variant Label"
          }
        },
        "required": [
          "product_id",
          "sku",
          "score",
          "similarity_signals"
        ],
        "title": "SubstituteResult",
        "type": "object"
      },
      "SubstitutesRequest": {
        "properties": {
          "filters": {
            "$ref": "#/components/schemas/RetrievalFilters"
          },
          "pos_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Accepted for client compatibility and ignored; scope comes from the token",
            "title": "Pos Id"
          },
          "product_id": {
            "minLength": 1,
            "title": "Product Id",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Why a substitute is needed, e.g. out of stock",
            "title": "Reason"
          },
          "top_k": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Top K",
            "type": "integer"
          }
        },
        "required": [
          "product_id"
        ],
        "title": "SubstitutesRequest",
        "type": "object"
      },
      "SubstitutesResponse": {
        "properties": {
          "candidates_returned": {
            "minimum": 0.0,
            "title": "Candidates Returned",
            "type": "integer"
          },
          "effective_pos_id": {
            "description": "Scope applied by the handler; always the token claim, never the body value",
            "title": "Effective Pos Id",
            "type": "string"
          },
          "low_confidence": {
            "default": false,
            "title": "Low Confidence",
            "type": "boolean"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/SubstituteResult"
            },
            "title": "Results",
            "type": "array"
          },
          "trace_id": {
            "description": "Correlation id, from the token claim when present",
            "title": "Trace Id",
            "type": "string"
          }
        },
        "required": [
          "trace_id",
          "effective_pos_id",
          "results",
          "candidates_returned"
        ],
        "title": "SubstitutesResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Internal HS256 service token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "jbg-ai",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/retrieval/products": {
      "post": {
        "description": "Over-fetch candidates so .NET can filter and still fill a page of `top_k`.\n\n`payload.pos_id` is ignored on purpose: the scope comes from the token.",
        "operationId": "retrieve_products_v1_retrieval_products_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetrievalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RetrievalResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid or missing internal service token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "Not implemented yet; delivered in a later change"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retrieve catalog candidates for a query",
        "tags": [
          "retrieval"
        ]
      }
    },
    "/v1/retrieval/substitutes": {
      "post": {
        "operationId": "retrieve_substitutes_v1_retrieval_substitutes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubstitutesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubstitutesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "Invalid or missing internal service token"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "description": "Not implemented yet; delivered in a later change"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Retrieve interchangeable products for a reference product",
        "tags": [
          "retrieval"
        ]
      }
    }
  }
}